# docs/openapi.yml
# THIS FILE IS GENERATED BY OpenApiSchemaExportTest — DO NOT EDIT MANUALLY.
# Run `./mvnw -pl service test -Dtest=OpenApiSchemaExportTest` to regenerate.
---
openapi: 3.1.0
components:
  schemas:
    Response:
      type: object
      properties:
        trace:
          $ref: "#/components/schemas/TelemetryTraceDto"
    Response1:
      type: object
      properties:
        spans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
    Response2:
      type: object
      properties:
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryMetricDto"
    Response3:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
    Response4:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryErrorGroupDto"
    SpanEvent:
      type: object
      properties:
        name:
          type: string
        epochNanos:
          type: integer
          format: int64
        attributes:
          type: object
          additionalProperties:
            type: string
        exception:
          type: boolean
    TelemetryErrorGroupDto:
      type: object
      properties:
        traceId:
          type: string
        serviceName:
          type: string
        errorSpans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
        errorLogs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
    TelemetryLogDto:
      type: object
      properties:
        epochNanos:
          type: integer
          format: int64
        severityNumber:
          type: integer
          format: int32
        severityText:
          type: string
        body:
          type: string
        traceId:
          type: string
        spanId:
          type: string
        serviceName:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
    TelemetryMetricDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        type:
          type: string
        value:
          type: number
          format: double
        epochNanos:
          type: integer
          format: int64
        serviceName:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
    TelemetrySpanDto:
      type: object
      properties:
        traceId:
          type: string
        spanId:
          type: string
        parentSpanId:
          type: string
        serviceName:
          type: string
        scopeName:
          type: string
        name:
          type: string
        kind:
          type: string
        startEpochNanos:
          type: integer
          format: int64
        durationMs:
          type: integer
          format: int64
        status:
          type: string
        statusMessage:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/SpanEvent"
    TelemetryTraceDto:
      type: object
      properties:
        traceId:
          type: string
        spans:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySpanDto"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryLogDto"
paths:
  /observability/api/telemetry/errors:
    get:
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response4"
      summary: Errors
      tags:
      - Workspace Telemetry Controller
  /observability/api/telemetry/logs:
    get:
      parameters:
      - name: query
        in: query
        schema:
          type: string
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: service
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response3"
      summary: Logs
      tags:
      - Workspace Telemetry Controller
  /observability/api/telemetry/metrics:
    get:
      parameters:
      - name: name
        in: query
        schema:
          type: string
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response2"
      summary: Metrics
      tags:
      - Workspace Telemetry Controller
  /observability/api/telemetry/slow-spans:
    get:
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: sinceMinutes
        in: query
        schema:
          type: integer
          format: int32
      - name: sort
        in: query
        schema:
          type: string
          default: duration
      - name: thresholdMs
        in: query
        schema:
          type: integer
          format: int64
          default: 500
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response1"
      summary: Slow Spans
      tags:
      - Workspace Telemetry Controller
  /observability/api/telemetry/traces/{traceId}:
    get:
      parameters:
      - name: traceId
        in: path
        required: true
        schema:
          type: string
      - name: repositoryId
        in: query
        schema:
          type: string
      - name: workspaceId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
      summary: Trace
      tags:
      - Workspace Telemetry Controller
info:
  title: qits-observability API
  version: 1.0.0-SNAPSHOT